🔧 Open up logging
Create methods on logger for add/remove handlers and get all handlers. Expose these in global methods as well on the global instance of LOG.

Open up the get methods for file and stream handlers

Remove the file logger since it now can be added afterward instead of by default.