Add a ciphertext field to the CiphertextElectionContest
Add a ciphertext field that is the elgamal accumulation of all of the selections for a contest.  This is the "contest total" in the specification that validators must check to verify the proof of selection limit.  While this value can be recomputed from available data, it is added to the output data set for convenience.