Add Guardian coefficient commitments to extended base hash
generate a commitment hash for guardians when generating the joint public key and propagate the commitment hash as part of a named tuple with the joint public key.  Use the commitment hash as part of the election builder and the cipertext election context, caching the result as part of the ciphertext election context.