Hash Guardian Commitments (#313)
* Add a ciphertext field to the CiphertextElectionContest

Add a ciphertext field that is the elgamal accumulation of all of the selections for a contest.  This is the "contest total" in the specification that validators must check to verify the proof of selection limit.  While this value can be recomputed from available data, it is added to the output data set for convenience.

* fix lint

* format

* Add Guardian coefficient commitments to extended base hash

generate a commitment hash for guardians when generating the joint public key and propagate the commitment hash as part of a named tuple with the joint public key.  Use the commitment hash as part of the election builder and the cipertext election context, caching the result as part of the ciphertext election context.

* fix linter errors

* Update the contest total ciphertext to be ciphertext_accumulation