🐛  Adjusting Spoiled Ballots and how they are handled (#305)
* 🐛  Adjust Spoiled Ballots to be PlaintextTallies

Spoiled Ballots are now PlaintextTally's that can be exported and imported. 
spoiled_ballots were removed from PlaintextTally to avoid circular reference. 
Tests were updated to ensure that the classes don't overlap. 
✅  Add test coverage


✅  Fix tests


✅  Fix test thats stuck


Remove unused methods that add to tests. 


Add announce into test for more coverage.

* Remove RSA function that's unused for coverage

* Refactor tally_selections to selections to match PlaintextTallyContest

* Add a generic CipherText Contest

Mimics the CiphertextSelection to allow some flexibility with decryption

* Remove Published Ciphertext Tally 

This only exists to handle the specific issue with a ciphertext tally containing spoiled ballots.

* Isolate Spoiled Ballots from Tally in Decryption Process

Spoiled ballots were intermingled with Tally and it trickled down into the lower classes. This isolates spoiled ballots to be its own decryption process even in the mediator. There is no big Tally file that includes the "tally" and the spoiled ballot collection.

* Fix Decryption Mediator Tests

* Fix tally tests

* Fix end to end election tests

* Disable Ciphertext Check awaiting decision

* Fix lint

* Add return for ballot box method

* Change Publish to continue to support PublishedCiphertextTally

* Remove larger loop functions for reconstruction of decryption shares

These force the import the use of DataStore in this file and are functions that would likely change for individuals using the functional methods. These are also no longer used by the decryption method which favorites the root class

* Remove unused import

* Changes to pull request

- Relocate code 
- Add comments