🔥. Remove publish
Publish functionally creates this idea that this method should be used when in reality the files should be handled individually by the system using it. This gives more flexibility on the output and less scaling problems with large ballot amounts. It also forces the use of serialization in the base electionguard.