♻️ Simplify Serialization (#394)
* 📄  Design and Architecture Shifts

- Remove focus on NamedTuple's in favor of dataclasses since serializers tend to handle these better
- Change name of section to

* ♻️. Prefer dataclass to NamedTuple

Using the built in constructor for dataclasses helps focus on plain constructable objects that are easy to serialize as opposed to immutable objects.

* 🚧  Remove Serializable


Squishy squash

* 📄. Update publish and verify documentation

- Publish and verify should be moved to spec documentation. Simplify wording for the python documentation.

* 📦 Remove jsons + Update dependencies

* 🔥. Remove publish

Publish functionally creates this idea that this method should be used when in reality the files should be handled individually by the system using it. This gives more flexibility on the output and less scaling problems with large ballot amounts. It also forces the use of serialization in the base electionguard.

* 🔥 Remove serialization test

* 🔨 Fix incorrect JSON data format

* 🐛 Fix nonce test bug

This doesn't happen all the time but is frustrating. Due to the setup of this test, it is possible for it to fail randomly. With a lower value of Q, it is more likely to happen more frequently. 

The issue arises due to sequencing. The separate random sequence CAN include the same number. It is random so sometimes there is a chance they align. Once this happens hypothesis keeps checking the case.

* 📦 Add pydantic to dev dependencies

* ♻️ Rename enum numbers to string

- This assists with typical serialization without specifying the output. This can also be accomplished with a dictionary in other languages
Revert "Squish thing"

This reverts commit e97efb6435520148328d757c0460a1ad68657df8.

* 🔥 Remove unnecessary prints in test_encrypt

* 🔨Have strategies rely on enum

Strategy can actual rely on the enum itself instead of a number

* 🐛 Fix extended data changes

* ♻️ Convert EncryptionDevice to dataclass

* ✨ Implement serialize

Add construct path to serialize

* ✅ Modify End to End Test

* ⚠️ Massive changes to group and q and p

* ✅ Documentation and Cleanup as per review